requirements.txt
pyTelegramBotAPI==4.15.4
supabase==2.3.4
python-dotenv==1.0.1
Flask==3.0.2
gunicorn==21.2.0